pushover: Handle more exceptions during setup
While looking into #110733, I noticed that when pushover fails to initialize
during a network outage, I get a `urllib3.exceptions.MaxRetryError`, whereas the
integration is only looking for ``BadAPIRequestError` or `ValueError`, leading
to a crash. If we instead handle the base `Exception`, we can log and retry
instead of crashing.

Fixes: #110733

Signed-off-by: Zack Cerza <zack@cerza.org>